fix(schedule): isolate validation from ambient state
